+++ home_applience_service_app_front_end/src/App.js
@@ -1,4 +1,4 @@
-import MainRouter from "./service/MainApplicationDriver/MainRouter";
+import MainRouter from "./service/MainRouter";
 
 function App() {
   return (
