+++ home_applience_service_app_front_end/src/service/MainRouter.jsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { AnimatePresence } from "framer-motion";
-import TechnicianDashboard from "../../Pages/Dashboards/TechnicianDashboard/TechnicianDashboard";
-import PageNotFound from "../../Pages/PageNotFound/PageNotFound";
-import PricingPage from "../../Pages/PricingPage/PricingPage";
-import AboutUs from "../../Pages/AboutPage/AboutUs";
-import CustLogin from "../../Pages/Login/CustomerLogin/CustLogin";
-import CustRegister from "../../Pages/Login/CustomerLogin/CustRegister";
-import ServicePage from "../../Pages/ServicesPage/ServicesPage";
-import UserSelection from "../../Pages/Login/UserSelection/UserSelection";
-import TechnicianLogin from "../../Pages/Login/TechnicianLogin/TechnicianLogin";
-import CustomerDashboard from "../../Pages/Dashboards/CustomerDashboard/CustomerDashboard";
-import HomePage from "../../Pages/HomePage/HomePage";
+import TechnicianDashboard from "../Pages/Dashboards/TechnicianDashboard/TechnicianDashboard";
+import PageNotFound from "../Pages/PageNotFound/PageNotFound";
+import PricingPage from "../Pages/PricingPage/PricingPage";
+import AboutUs from "../Pages/AboutPage/AboutUs";
+import CustLogin from "../Pages/Login/CustomerLogin/CustLogin";
+import CustRegister from "../Pages/Login/CustomerLogin/CustRegister";
+import ServicePage from "../Pages/ServicesPage/ServicesPage";
+import UserSelection from "../Pages/Login/UserSelection/UserSelection";
+import TechnicianLogin from "../Pages/Login/TechnicianLogin/TechnicianLogin";
+import CustomerDashboard from "../Pages/Dashboards/CustomerDashboard/CustomerDashboard";
+import HomePage from "../Pages/HomePage/HomePage";
 
 function MainRouter() {
   return (
